Changed how bk filename is obtained. Added HU analysis type logic. Changed huHazardMapWindSpeed to huRPWindSpeeds.
Added functions to get the bkfilename from the hpr comment to resolve any issue when there are multiple bk files that the correct bk file is used. Added Hurricane analysis type logic and stubs for flood and tsunami. Added else to geographic unit and count to avoid any python variable called before assignment. Changed huHazardMapWindSpeed to huRPWindSpeeds as this the huhazardmapwindspeed is for the toc and the rp is for the returnperiod. Added hurricaneTrack function stub in.